Formalize and expose an API for usage as a library.
