+++ build.gradle
@@ -32,7 +32,7 @@ task fatJar(type: Jar, dependsOn:'compileJava') {
   classifier 'fat'
 
   manifest {
-    attributes 'Main-Class': 'com.jakewharton.dexmethodlist.Main'
+    attributes 'Main-Class': 'com.jakewharton.dex.DexMethods'
   }
 }
